+++ CustomLoggers/DatabaseLogger/DbLoggerOptions.cs
@@ -4,4 +4,5 @@ public class DbLoggerOptions
 {
     public required LoggerDbContextFactory Factory { get; set; }
     public required LogLevel MinimumLogLevel { get; set; }
+    public required IHostEnvironment Environment { get; set; }
 }